requirements.txt
PyMuPDF==1.23.8
nltk==3.8.1
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
streamlit==1.31.0
fpdf2==2.7.8
tqdm==4.66.1
